Time: 42 ms (44.57%), Space: 31.1 MB (87.76%) - LeetHub

+++ 724-find-pivot-index/724-find-pivot-index.cpp
@@ -3,9 +3,9 @@ public:
     int pivotIndex(vector<int>& nums)
     {
         int ret = -1;
-        long sum = 0;
+        int sum = 0;
         sum = accumulate(nums.begin(),nums.end(),sum);
-        long left_sum = 0;
+        int left_sum = 0;
         
         for(int i=0;i<nums.size();++i)
         {
